Move edge label helper
