+++ lib/formatEdgeLabel.ts
@@ -1,0 +1,10 @@
+export function formatEdgeLabel(
+  depName: string,
+  requiredRange: string,
+  latestVersion: string,
+  isLatest: boolean,
+): string {
+  return isLatest
+    ? requiredRange
+    : `${depName}\n${requiredRange} / ${latestVersion}`
+}
